% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratum_to.R
\name{helper functions (previous/following stratum)}
\alias{helper functions (previous/following stratum)}
\alias{stratum_to}
\alias{stratum_from}
\title{Two helpers to select stratum connected to a flow}
\usage{
stratum_to(x = .data, value = "stratum")

stratum_from(x, value = "stratum")
}
\arguments{
\item{x}{usually the .data pronoun}

\item{value}{="stratum", the value that should be returned, e.g. "stratum" or "x"}
}
\value{
value mapped to the left (from) or right (to) side of the alluvium.
}
\description{
Two helpers to select stratum connected to a flow
}
\examples{
# stratum_to(.data)
# stratum_from(.data)
}
