% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_flow_interactive.R
\name{geom_flow_interactive}
\alias{geom_flow_interactive}
\title{Flows between lodes or strata}
\usage{
geom_flow_interactive(...)
}
\arguments{
\item{...}{arguments passed to base function,
plus any of the interactive_parameters.}
}
\description{
This geometry is based on \code{\link[ggalluvial:geom_flow]{ggalluvial::geom_flow()}}.
See the documentation for those functions for more details.
}
\examples{
# add interactive contours to a ggplot -------
library(ggplot2)
library(ggiraph)
#...
#x <- girafe(ggobj = p)
if (interactive()) print(x)
}
\seealso{
\code{\link[ggiraph:girafe]{ggiraph::girafe()}}
}
