% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_flow_interactive.R
\docType{data}
\name{GeomInteractiveFlow}
\alias{GeomInteractiveFlow}
\title{ggproto class for ggiraph}
\description{
ggproto class for ggiraph.

geom_flow_interactive returns a layer that contains a GeomInteractiveFlow
object. The GeomFlow object is responsible for rendering the data in the plot.
}
\keyword{datasets}
