% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_stratum_interactive.R
\docType{data}
\name{GeomInteractiveStratum}
\alias{GeomInteractiveStratum}
\title{ggproto class for ggiraph}
\description{
ggproto class for ggiraph.

geom_stratum_interactive returns a layer that contains a GeomInteractiveStratum
object. The GeomStratum object is responsible for rendering the data in the plot.
}
\keyword{datasets}
